fix nested typechain from cp command
