+++ packages/types/src/contract/typechain/captcha/types-arguments/captcha.ts
@@ -1,17 +1,36 @@
 import type BN from 'bn.js';
 
-export type AccountId = string | number[]
-
 export type Hash = string | number[]
+
+export type AccountId = string | number[]
 
 export enum LangError {
 	couldNotReadInput = 'CouldNotReadInput'
 }
 
+export enum Payee {
+	provider = 'Provider',
+	dapp = 'Dapp'
+}
+
+export enum DappPayee {
+	provider = 'Provider',
+	dapp = 'Dapp',
+	any = 'Any'
+}
+
+export enum GovernanceStatus {
+	active = 'Active',
+	inactive = 'Inactive'
+}
+
 export enum Error {
-	notAuthorised = 'NotAuthorised',
+	notAdmin = 'NotAdmin',
+	notOwner = 'NotOwner',
 	contractTransferFailed = 'ContractTransferFailed',
+	providerAccountExists = 'ProviderAccountExists',
 	providerExists = 'ProviderExists',
+	providerAccountDoesNotExist = 'ProviderAccountDoesNotExist',
 	providerDoesNotExist = 'ProviderDoesNotExist',
 	providerInsufficientFunds = 'ProviderInsufficientFunds',
 	providerInactive = 'ProviderInactive',
@@ -33,59 +52,7 @@ export enum Error {
 	noCorrectCaptcha = 'NoCorrectCaptcha',
 	notEnoughActiveProviders = 'NotEnoughActiveProviders',
 	providerFeeTooHigh = 'ProviderFeeTooHigh',
-	commitAlreadyExists = 'CommitAlreadyExists',
-	captchaSolutionCommitmentAlreadyExists = 'CaptchaSolutionCommitmentAlreadyExists',
-	verifyFailed = 'VerifyFailed'
-}
-
-export enum Payee {
-	provider = 'Provider',
-	dapp = 'Dapp'
-}
-
-export enum DappPayee {
-	provider = 'Provider',
-	dapp = 'Dapp',
-	any = 'Any'
-}
-
-export enum GovernanceStatus {
-	active = 'Active',
-	inactive = 'Inactive'
-}
-
-export type Commit = {
-	id: Hash,
-	user: AccountId,
-	datasetId: Hash,
-	status: CaptchaStatus,
-	dapp: AccountId,
-	provider: AccountId,
-	requestedAt: (number | string | BN),
-	completedAt: (number | string | BN),
-	userSignaturePart1: Array<(number | string | BN)>,
-	userSignaturePart2: Array<(number | string | BN)>
-}
-
-export enum CaptchaStatus {
-	pending = 'Pending',
-	approved = 'Approved',
-	disapproved = 'Disapproved'
-}
-
-export type LastCorrectCaptcha = {
-	before: (number | string | BN),
-	dappId: AccountId
-}
-
-export type CaptchaData = {
-	provider: AccountId,
-	datasetId: Hash,
-	datasetIdContent: Hash
-}
-
-export type User = {
-	history: Array<Hash>
+	commitAlreadyExists = 'CommitAlreadyExists'
 }
 
 export type Provider = {
@@ -105,10 +72,49 @@ export type Dapp = {
 	payee: DappPayee
 }
 
-export type RandomProvider = {
-	providerId: AccountId,
-	provider: Provider,
-	blockNumber: (number | string | BN),
+export type UserHistorySummary = {
+	correct: (number | string | BN),
+	incorrect: (number | string | BN),
+	score: (number | string | BN)
+}
+
+export type Commit = {
+	id: Hash,
+	userAccount: AccountId,
+	datasetId: Hash,
+	status: CaptchaStatus,
+	dappContract: AccountId,
+	providerAccount: AccountId,
+	requestedAt: (number | string | BN),
+	completedAt: (number | string | BN),
+	userSignaturePart1: Array<(number | string | BN)>,
+	userSignaturePart2: Array<(number | string | BN)>
+}
+
+export enum CaptchaStatus {
+	pending = 'Pending',
+	approved = 'Approved',
+	disapproved = 'Disapproved'
+}
+
+export type LastCorrectCaptcha = {
+	before: (number | string | BN),
+	dappContract: AccountId
+}
+
+export type CaptchaData = {
+	providerAccount: AccountId,
+	datasetId: Hash,
 	datasetIdContent: Hash
+}
+
+export type User = {
+	history: Array<Hash>
+}
+
+export type RandomProvider = {
+	providerAccount: AccountId,
+	provider: Provider,
+	blockNumber: (number | string | BN)
 }
